M3-R phases 0 and 1: correct the record, repair the instruments
Phase 0 corrects what the verification retracted. The '4.6%' comparison is gone
from all four documents and the registry; M4's Blocked reason now names the two
things actually blocking it rather than the withdrawn diagnosis.

Phase 1 rebuilds the three instruments the M3 verdict rested on.

The collapse detector now reads audio. It read row 0 only, which is why
m3-collapse.json records the control as silent_count 0 while 17 of its 30
general30 generations emit the Mimi silence token on up to 113 of 124 frames.
Thresholds come from a calibration file that states plainly it was frozen after
the candidates existed, and shows the verdict is flat across the plausible range.

Condition 4 gets a calibration band and an interval. The old bar was 'higher on
8 of 10', whose power at a true win rate of 0.70 is 0.383 - it could not tell an
absent effect from an unmeasured one. The band is the target speaker's own
leave-one-out spread, mean 0.8166, and the bar is now a fraction of the distance
from the control to it.

Training loss detail no longer depends on W&B, and generate.py requires a seed.
Every M3 count was a single unrecorded draw.

Two fixes came out of verifying the above rather than from the plan.

DEFAULT_MIN_BAND_CLOSURE could be moved from 0.25 to 0.05 with all 473 tests
still passing, and it decides live arms. Pinned, with a mutation test.

experiment_budget.py held HARD_CAP = 100.0 after the cap moved to 125, so it
returned non-zero for every possible plan - the one instrument that can say
'stop' had stopped meaning anything. The thresholds are now fractions of a cap
the caller passes. Under it, phase 4's full 6-hour plan is rejected: US.04
predicted against a US.50 new-run limit. That verdict is recorded as a test
so nobody weakens the bar to admit the plan.

311 tests to 479.

+++ tools/tokenize_text.py
@@ -263,7 +263,14 @@ if __name__ == "__main__":
     )
     parser.add_argument(
         "--audio_tokenizer_frame_rate",
-        type=int,
+        # float, not int. Mimi runs at 12.5 Hz, which is what the default already says, and
+        # tokenize_and_pad_text divides by this to place every token in a frame. Under
+        # type=int, `--audio_tokenizer_frame_rate 12.5` raises ValueError and `12` is
+        # accepted silently - and a silent 12 slides the text stream progressively out of
+        # step with the audio, by one frame every two seconds, with nothing to see in any
+        # artifact. The default is a float literal, so argparse never coerced it and the
+        # shipped M3 datasets are unaffected (reports/m3-text-stream-audit.json).
+        type=float,
         default=12.5,
         help="Frame rate for the audio tokenizer.",
     )
